add: add emptylocker test for bucket

+++ example/main.go
@@ -46,7 +46,7 @@ func main() {
 
 	options := cache.DefaultOptions
 
-	for _, arg := range []uint8{3, 5, 10} {
+	for _, arg := range []int{3, 5, 10} {
 		options.EvictInterval = arg
 		options.ConcurrencySafe = false
 		fmt.Println("=====Options=====")
